requirements.txt
torch==2.3.0
torch-geometric==2.5.3
torch_scatter==2.1.2
torch_sparse==0.6.18
scikit-learn==1.5.0
numpy==1.26.4
pandas==2.2.2
PyYAML==6.0.1
ruff==0.4.8
pytest==8.2.2
